requirements.txt
psutil~=5.9.8
asyncer~=0.0.7
numpy~=1.24.3
sympy~=1.12
websockets~=11.0.3
pytest~=7.3.1
fastapi~=0.111.0
aiohttp~=3.9.5
uvicorn~=0.29.0
loguru~=0.7.0
boto3~=1.28.33
botocore~=1.31.33
joblib~=1.2.0
pandas~=2.0.2
oss2~=2.18.6
DBUtils~=3.0.3
PyMySQL~=1.1.1
oracledb~=1.3.1
pymssql~=2.3.0
psycopg2~=2.9.9
PyJWT~=2.7.0
python-dateutil~=2.8.2
cyksuid~=2.0.2
jieba~=0.42.1
torch~=2.1.0
tqdm~=4.65.0
statsmodels~=0.14.0
scikit-learn~=1.2.2
scipy~=1.10.1
lightgbm~=4.1.0
Cython~=3.0.10
setuptools~=70.0.0
rsa~=4.9
prophet~=1.1.4
pydantic~=2.7.1